+++ tests/colorful_print/colorful_print_various_words_test.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+# shellcheck source=/dev/null
+source "$(dirname "${BASH_SOURCE[0]}")/../../hooks_utility.sh"
+
+hooks_utility_colorful_print_pass
+printf '%b' "\n"
+
+hooks_utility_colorful_print_fail
+printf '%b' "\n"